Removed scratch work.
Decided not to use this approach.  Letting the trailing ellipsis come
after the list is much easier to describe than trying to make it part of
the last element in the list.  The number of commas is still one less
than the number of elements, as it should be -- the ellipsis doesn't
become an element even though it's a separate token.