Shortening of some lines and variable names to fix swifttest errors.
Turns out that the REPL can't cope with wrapped lines. In some cases I've fixed this by shortening the names of functions or variables; in others, the code is simply too long to fit in an 80-character limit without line-wrapping, and so I've left it overflowing. (It's still copy-and-pasteable.) I figure there's no point fixing this until we know the properties of the actual environment (REPL, playground etc.) we'll be using.