Renamed named-expression to identifier-expression.
Per discussion with Brian, it's an identifier used as an expression in
the context, not an expression that has a name.  The value of the
identifier has a name, and you set the value of the identifier using
assignment from an expression.  But without quoting/delay/promise, it's
not the expression itself that is being named -- rather the value that
it evaluated to.