Removed the old Language Tour testcode directives.
The old Language Tour is constantly getting out of date, and isn't linked to from anywhere within the Language Guide. I've removed its testcode hooks so that I can test the Guide as a whole and trust that the number of reported errors is representative of the parts of the book that are actually intended for publication.