Settled on calling these "prefix expression".
Using the same name for the section head, the term being discussed, and
the grammatical category -- just like we do elsewhere.  Especially in
the heading, it makes more sense to call this an expression than an
operator.  The prefix operator is what turns an expression into a prefix
expression, but the section is about more than just the operators.