Moved "Deinitializers" to the end of the Initialization chapter.
It feels appropriate that the last thing you do should be at the end of the chapter.
Also, in an ARC world, they'll probably be reasonably rare.
Putting them at the end of the chapter makes an otherwise hefty section eminently skippable if you think "I don't need one of those right now".