Removed only-child heading, promoting its children.
This is purely a style issue -- I don't want only one heading at a given
level with multiple headings under it.  The other alternative would have
been to promote it to become a sibling to prefix/infix/postfix
expressions, but that doesn't seem right: the built-in operators being
described are all binary/infix operators.