First draft of a section on Arrays for Basic Types.
The Array implementation is still in flux, but it's a large hole in the documentation at the moment, and I wanted to get a sense of the size of it with the Guided Tour in mind.
Based on this first draft, I think "Arrays and Dictionaries" might be better off as a new separate chapter.