Omitted assignments when discussing closures.
Bare closure expressions aren't valid at the REPL (or anywhere else),
which means you have to put them in a different context to start playing
with them.  Leaving off the context here makes the listing much shorter
and removes the need for type annotations, which are not at all the
topic of the current discusion.