Extensive rewrite of the Functions chapter to describe the new unified function syntax, which now works* ! Also switched over to saying "definition" rather than "declaration" in this chapter.
* Well, I say it works… there's currently a problem with the local parameter names,
in that you have to use the *local* name when you call the function, not the external name.
This is Not Right, and will be fixed by Doug very shortly.
In the meantime, there are several code lines herein marked as FIXME which will need reinstating once the fix is made.