Added note about bare expressions in PG.
This text follows the first time this trick is used, and should help
clear up any confusion about why we'd do this.  Writing a bare
expression in normal code doesn't make any sense, but in a playground
it's a nice trick to check the value of a variable or other expression.

Fixes <rdar://problem/17190948> Introduce the notion of writing a bare
expression in a playground