Updated expectation with new array append behavior.
The += operator now expects you to write result += [item] which feels
like it's a bit confusing (a bit too much syntax) in the tour.  It's not
as obvious to a novice that [item] is just an array literal for a
one-element array -- it kind of looks like special magic.  Using
.append() instead is cleaner.