Performed a fairly comprehensive rewrite of the Inheritance chapter in light of the seed 5 "override" changes.
This rewrite also fixes a number of other problems with this chapter. These include the fact that description was a method rather than a property,
and the fact that the example we used for getter / setter overrides wasn't really best practice coding style.
This chapter no longer mentions initializers at all, which is a much-needed structural improvement for the overall book.