Removed the Guide's description of allowed custom operator characters.
This list is defined completely in the Reference, which also includes the complex rules around the use of period characters, and rules about whitespace characters.
There's no point defining the character list (and its rules) incompletely in the Guide as well - we may as well make the Reference be the canonical version.