Backed out incorrect prose around 'try'.
Reverts some changes from 613c72e.

In the guide we talk about try being used with functions because that's
the most common use case, but any expression can go in a try expression,
and try expressions themselves are expressions.  This is confirmed by
reading that part of the parser and agrees with the whitepapers and
emails from the compiler team.