Copy stable links table from Xcode 7 Beta 3 (3dde2cb).
The following stable IDs appear in the content but didn't get added to
the JSON file:

* collectiontypes-creatinganemptyarray
* collectiontypes-creatinganarraywithadefaultvalue
* collectiontypes-creatinganarraybyaddingtwoarraystogether
* collectiontypes-creatingandinitializinganemptyset
* collectiontypes-creatingasetwithanarrayliteral
* collectiontypes-fundamentalsetoperations
* collectiontypes-setmembershipandequality
* collectiontypes-creatingadictionarywithadictionaryliteral

That fix probably should have happened on 'main' rather than on a seed
branch, but in the run-up to WWDC it got done over there instead.
Adding these stable link table entries on 'main' means we can build
[Tool S] XML on this branch, and also prevents merge issues if we happen to
add a section to 'main' before merging in the Xcode 7 beta branch.