Clean up discussion of inferred closure types.
See feedback 274828.  Some of the issue here was leftover mention of the
second argument of the sort() function which no longer exists.

Considered adding a cross reference to the Generics chapter, but I don't
think there's a natural spot for it, and the discussion here doesn't
seem to lead towards the reader jumping way ahead in the book to read
about generics, even though that is an important part of how the type
inference around sort() happens.