WIP: Another approach for the code listing.
This one has the advantage of following on from the scenario used in the
listings before it of customers waiting in a line.  However, it's rather
more complicated -- the use of a closure is contrived and I end up with
an inout parameter and explicit array indices.  I suspect it will be
even more contrived when it comes to writing an escaping version.
diff --git a/docs/canonical/LanguageGuide/Closures.rst
b/docs/canonical/LanguageGuide/Closures.rst index f5b6c2c..a4cadd3
100644 --- a/docs/canonical/LanguageGuide/Closures.rst +++
b/docs/canonical/LanguageGuide/Closures.rst @@ -692,6 +692,37 @@ Here's
a simplified version of that function: -> assert(true, message: {"This
closure is run."}) <- This closure is run.

+
+
+
+For example, the ``serveNextCustomer(_:counters:)`` function in the listing below
+takes a customer's name and a list of service counters.
+Service counters that are available have a ``nil`` value
+and counters that are in use have the name of the customer
+who is being helped at that counter.
+The customer's name is passed to this function wrapped in a closure
+and is evaluated only if there is an available counter.
+
+::
+
+    var customersInLine = ["Chris", "Alex", "Ewa", "Barry", "Daniella"]
+    var counters: [String?] = [nil, nil]
+
+    func serveNextCustomer(customer: () -> String, inout counters: [String?]) {
+        for i in counters.indices {
+            if counters[i] == nil {
+                counters[i] = customer()
+                return
+            }
+        }
+    }
+
+    serveNextCustomer({customersInLine.removeAtIndex(0)}, counters: &counters)
+
+
+
+
+
 You can use the ``autoclosure`` attribute on the function parameter,
 which indicates that the expression being passed
 should be automatically wrapped in a closure.