Remove all discussion of curried functions.
(Per discussion with Brian)  The reason that the special syntax was that
it didn't get enough use to justify its complexity.  We don't include
conceptual information in the reference, so if we wanted to keep this
material it would have to move to the guide.  And it doesn't make sense
to put in the guide specifically because it's so seldom used.