Backticks are needed for var/let/itout on parameter.
This follows the typical pattern of meaning, error, new meaning.  In
Swift 2, these were valid and had their old meaning.  Before we give
them a new meaning --- func f(inout x: Int) means "inout" is the
argument label for the parameter --- we make them an error for a major
release cycle.  So the assertion is expected to start failing sometime
in Swift 4.