Unlabel first argument of delegate methods.
This isn't covered in the API guidelines, but they don't read well with
an argument label:

    gameDidStart(game: foo)
    gameDidStart(foo)

This is the typical Cocoa naming convention for delegates.  It might
work to move "game" from the method name to an argument label as in
"didStart(game:)" but that pattern doesn't work for the new-turn method.