Frame in terms of what you do, not what type it is.
The 'indices' property of String.CharacterView returns a value of type
DefaultBidirectionalIndices<String.CharacterView> -- but what actually
matters from the API contract is that any collection has an 'indices'
property which returns a sequence of type Self.Indices.  (Self.Indices
has to conform to IndexableBase and Sequence, but not Collection.)

Per feedback from Brian.