Reference the detailed designated initializer inheritance rules.
<rdar://problem/23137596> TSPL: guide/reference mismatch about inheriting designated initializers

Previously, the description here didn't include some caveats about the inheritance
rules. Since the rules are somewhat complex and detailed in the language guide,
I put a reference to that description rather than duplicating the detail definition.

The next few existing sentences in the reference have a little overlap with the
language guide, but for now I'm considering that to be okay.