Remove references to the Boolean  protocol.
Fixes: <rdar://problem/27435106> TSPL: SE-0109 Remove the Boolean protocol

This would be really straightforward except for issues related to external
Boolean types which are bridged manually in the compiler.
(DarwinBoolean, ObjCBool, CBool).

Alternatives considered:
(1)
The value of the *condition* must have a type that converts to
the ``Bool`` type.

(2)
The value of the *condition* must have a type that converts to ``Bool``.

(3) <-- runner up
The value of the *condition* must be of type ``Bool``.

(4) <-- bad because the guide claims implicit conversions aren't in Swift
The value of the *condition* must be implicitly convertible to ``Bool``.

(5)
The value of the *condition*
must be of type ``Bool``
or one of the bridged Boolean types ``DarwinBoolean``,``ObjCBool``, or
``CBool``.

(6) <-- in this commit
The value of the *condition*
must be of type ``Bool`` or a type bridged to ``Bool``.