Remove @noreturn and describe the Never type.
WIP. The description of Never is a bit shorter than the previous combined
docs for @noreturn as a type attribute and as a declaration attribute.
That's partially because of duplication, and partially because Never fits into
the type system in a more coherent way than before, so it's easier to explain.

Not sure if an example is needed. The obvious choices are an infinite loop and
a fatalError(), but those are somewhat degenerate cases. An infinite loop example
could be expanded a bit to, say, grab a number from the REPL and always multiply
it by 2, or something, to indicate that a really simple perpetual program could
use something like Never to repeat some simple behavior forever.

Fixes <rdar://problem/27433854> TSPL: SE-0102 Remove @noreturn attribute and introduce an empty NoReturn type