Defer addition of a section.
Per discussion with Brian and Kyle, this section needs additional work
before we ship it.  The new material is showing a common operation that
many (most) value types should be doing, but it's not strictly speaking
new from the SE proposal.  The way you defined Equatable stays the same,
but what changes is that the way you conform to it is no longer the odd
"add stuff at global scope and an empty protocol conformance" approach.

Filed <rdar://problem/27556117> TSPL: Revise the discussion of Equatable,
Comparable, and operator requirements in protocols