Don't give "&" a name.
We don't have precedent for operators that act on types.  It's easier to
come back later and give it a name, if we decide that this really is the
right thing to do, than it is to remove a name that's already been
published.

Per review with Brian and Kyle.