Remove a misleading clause.
Because we called out just structs and enums specifically, people thought this
meant that the behavior didn't apply to classes. It does.

The alternative would be to add 'and classes', but the whole section is talking
about all three anyway.

Fixes <rdar://problem/27381426> TSPL: Methods - misleading comment about structs/enums but not classes