Rename dynamicType to type(of:).
Changes the docs to reflect that dynamicType was renamed to type(of:) in
SE-0096.

I kept the parts that refer to it as a "dynamic type expression", because it's
still an expression and still fetches the dynamic type specifically.

I removed `dynamicType` from the list of lexical keywords, but didn't add in
`type`, because you *can* use it as an identifier. (Not true of `dynamicType`.)

<rdar://problem/26757527> TSPL: SE-0096 Convert "dynamicType" from a property to an operator