Add a discussion of non-returning funcs to the Functions chapter.
Addresses <rdar://problem/27433854>
TSPL: SE-0102 Remove @noreturn attribute and introduce an empty NoReturn type

This commit adds a sub(sub?)section in Functions Without Return Values with
a description of Never and non-returning functions.

We need to talk about Never somewhere, and it's not appropriate to talk about
it in § Attributes anymore. In previous commits I experimented with putting a
section in § Types, but that ended up not working because all of the other
types have a grammatical component to them.