Order if/else-if correctly.
This code was added in 15b0c3a3, but in the wrong place.  As originally
written, the else-if clause is never called -- a non-nil value for
buildingName is necessary for the second conditional but sufficient for
the first conditional.  This means any time buildingName is non-nil, the
first branch is taken.

See feedback 293594.