Fix example of examining a dynamic type.
The === comparison appears to work because it returns false when you
expect it to, but it never returns true.  Why?  Because it's comparing a
class against an instance, which is never true.  The type(of:) call on
the left returns a value of type SomeBaseClass.Type, but the
postfix-self expression someInstance.self expression on the right hand
side is the same assomeInstance, which is of type SomeBaseClass.

There isn't a way to spell "the static type of foo" like foo.self was
trying to do.  For now, fix the listing by using an actual/hard-coded
type in the rhs of the == comparison.