Incorporate more Doug feedback.
Per discussion in person with Brian and Kyle, although we like the
parallelism of showing the syntax of all three forms of extension here,
the reader is much more likely to be confused by the generic form.  A
very brief cross reference with a discussion of the goal --
conditionally adding functionality -- is a better way to meet the
reader's needs.