Add a proposed outline for the chapter.
Start with the common case -- structures -- and walk through some common
objections.  Then get into the fact that classes offer more powerful and
dynamic behavior, but that it comes at a cost to programmer cognitive
load and to performance.

From discussion with [Contributor 4485], this looks like it will work better.