Remove leftover prose, add to-do comments.
I don't want to discuss raw memory or memory addresses here because they
don't really help if you're a beginning programmer who has never had to
think in pointers.

I'm also avoiding the discussion of windows because it's really just an
AppKit and UIKit artifact from being designed and implemented in
Objective-C.  If you haven't worked in those systems, you might think
that GUI code in general is an exception to the "class root with struct
leaves" model described earlier.  It's not, but that's not the point
here.

I've filed <rdar://problem/31490238> to track adding discussion of
performance differences between classes and structs.