Rough in one approach to the new key path grammar.
Depending on the number of special cases, we may end up having to just
model this as type-dot-expression and handle the overproduction in
prose.  Using an explicit member expression for the part that starts
with the dot would remove a bunch of things we want to exclude -- binary
expressions, assignment, try expressions, and closure expressions.