Remove indirection through characters property.
You can iterate over the character view directly by iterating over the
string itself -- under the new model, String is a collection again.
We may also want to restore the example of a for loop iterating over a
string, which was deleted in 57ed66e56206161e2d9427479fb399502e982dd2.