Fix warnings about unused values.
First preference is to just use the value, assuming that's idiomatic and
fits the flow of the example.

Using the hidden "_ = foo" pattern to silence the warning in places
where the body of the if-let is intentionally blank, to be filled in by
other code.

An alternate fix for 'kind' in the tour would have been to change
(kind, numbers) to (_, numbers) but that's not as approachable.  Using
underscore to ignore something doesn't come up properly until later in
the tour.