Incorporate feedback from Brian on the reference.
The grammar for multiline quoted text could also be written as:

    multiline-quoted-text --> multiline-quoted-text-item multiline-quoted-text-OPT
    multiline-quoted-text-item --> quoted-text-item
    multiline-quoted-text-item --> The Unicode scalar value ``"``, U+000A, or U+000D

But I think that's a little less clear.  You have to read both
quoted-text-item and multiline-quoted-text-item to see that the former
doesn't allow " CR and LF but the latter does, so those three characters
are allowed in a multiline string literal.