Re-order for better flow.
Move the "friendly" bits back to the beginning, as they were on
'develop', and move the "industrial quality" bit back to being connected
to "friendly".

The bullets are solutions, not problems, so flip the sentence before
them to introduce them correctly.

Remove one of several "Swift is fast" statements.  There are enough of
these already.

Per in-person revision with Brian.