Mention the Obj-C prefix.
This isn't new, but we never pointed out this
potential tripping point in the past.  For @objc() on class members,
there's no need for a prefix, because they just need to be unique with
in the class.  But classes don't have their Swift module prefix on their
manual Obj-C names, so you have to prefix them.

Per comments from [Contributor 6004].