Correct our statement of the model.
This matches what's described in the SE proposal:

    For purposes of access control, extensions to any given type T
    within a file are considered to be a single access control scope,
    and if T is defined within the file, the extensions use the same
    access control scope as T.

As in the SE proposal, the desired use case is extensions being able to
access private members of types they extend.  So I'm not devoting
discussion to the behavior of two extensions to the same type that are
both in the same file, but not in the file where the type is defined.

Follow-on to 97fb830bd6fd030a492fea574eb53b8b44964de1, with more tech
review from [Contributor 6004].