Show nonexclusive-but-safe using a global vs a local.
This avoids the need to bring up a complex code listing, with a design
pattern (private stored variable backing a public computed variable)
that hasn't been shown in detail elsewhere in the book.  It also brings
the list of conditions that have to be true for the compiler to prove
such access is safe into the main flow.

Per some discussion this afternoon with Brian.