Address some old feedback from Brian.
Better names for a couple functions & variables.  The API naming
guidelines suggest the -ed/-ing rule when you have a
mutating/nonmutating pair of methods, but that guidance doesn't apply
well to functions.  We don't use the "inPlace" suffix anywhere else in
the book, and it's a pattern they avoided in the standard library, so
don't use it here.

It's really odd to have a variable named "one" whose value is 2.

Tie the arbitrary myNumber/myOtherNumber variables forward to the game
examples by making them players' scores.

Various copyediting.