Incorporate feedback from Brian.
Restructure for better flow in early segments -- all of the
multithreading comments are together, and the "what is an access?" bit
fits better back at the beginning.

Using the name "Holly" from a marketing image of the Apple Watch.  Looked
at examples we had in the original code listings, but Vi and Valencia
seem potentially troublesome.  Emacs users might get grouchy about the
former and non-Catalan speakers might read the latter as special
treatment.

Took a pass over all of the places we call out a conflict or lack of a
conflict to use consistant wording.

Took another pass over "// Error" usage, adding a terse explanation of
why it's an error.