Remove FIXME
I don’t think we need to make this comparison, since the two
implicit behaviors are different enough that it would take
effort to describe their similarity/difference. In addition, the
compiler is still having enough trouble with type inference that
the slightly hand-wavy “in contexts where type inference can
determine the implied type” is really all we can say about this.