Group property wrappers with observers & computed.
Per discussion with Chuck.  Conceptually, the topics around what you can
do with a property fit together, and the discussion of locals and
globals and type properties form a sort of "additional topics" chunk at
the end.  Property wrappers naturally follow on from computed properties
because a wrapper is implemented as a computed property, and wrappers
are a way to avoid repeating the same code in a bunch of computed
properties & property observers.