Rough in an example of SwiftUI-like syntax.
As discussed in the review with Brian and Kyle today, the examples
you've seen so far don't looks like  SwiftUI code.  The existing
examples don't ever access a wrapped property or its projection from
within the scope of that type, so you never have an implicit/explicit
self in those examples.