Incorporate tech review from [Contributor 6004].
The symbol name in the binary, as seen by 'nm', can also be impacted by
the choice of argument to @objc -- but the compiler may or may not even
generate a symbol for a given class.  The Obj-C runtime name is the one
that developers are most likely to care about.