Separate the two selector expression examples.
When they were one big example in the REPL, it was fine -- at the
point we referred to doSomething(_:) in the first example, it wasn't
ambiguous because the extension hadn't been read in yet.  That's not the
case now that we're compiling tests.

Also mocked up a version of the second listing that redefined SomeClass
instead of extending it, but that ends up just being noisy.