Remove discussion of hashValue.
At the time this was originally written, we likely didn't have a full
discussion of how to adopt Hashable somewhere else, so it ended up here.
That's no longer the case today -- the Hashable reference has a section
with all the details, and that's the correct document to "own" that
information.