More live editing with Brian and Chuck.
No need to mark the rationale for passive, since it will be the same
everywhere that we need it in reference.

Reword a subheading that could be misread as a more general top-level
heading.

Reserve "transform" for the action that the compiler and the language
feature do, to turn the DSL into calls to the result builder.