Incorporate feedback from Brian.
The list of result-building methods really isn't a protocol, although
it's similar to an Objective-C informal protocol.  They're methods with
special names that give you special behavior when you implement them.

Don't mention of DSLs for now.  If we come back later and frame result
builders as one of Swift's options for building DSLs, we can add back a
more substantial form of that narrative.

Use triple-hyphen instead of a literal em-dash, matching what we do
everywhere else.  (This was probably a copy/paste error.)  In
monospaced text, it's almost impossible to tell apart hyphens and
em-dashes.

Elsewhere in the reference, we call them "while statements", not
"while loops", so match that here too.