Fix up tested code.
The deprecation of runAsyncAndBlock without an agreed-upon replacement
is a known issue.  Not a big deal here because it's used only in hidden
setup lines.  Not sure why these were originally marked as needing to be
compiled -- maybe it was a leftover from when I was getting async to
work at all.