Add an example of the if-let sugar to the tour.
Thematically, it fits a little better after the initial if-let example,
but up there the optional value is called 'optionalString' to make
things clearer the first time you see an optional. In a case like that, you
wouldn't actually want 'optionalString' to be the name of the
non-optional unwrapped value.

Putting it after the '??' example instead makes 'nickname' available,
which is a sensible constant name for an optional or non-optional.  I
think this flow works: introduce 'if'-'let', call out '??' as an
alternative, show the (slightly more advanced) syntactic sugar.