Use a more approachable syntax for ignored errors.
Stopping the DocC preview produces a non-zero exit status [1], which is
understood by make(1) as a failed build -- but in this case, it's
expected and shouldn't cause make(1) to give up or report an error.

The prefix dash ignores a non-zero exit from only the command that it
marks, which is more specific and usually more correct.  However, a
prefix hyphen looks like a typo to those less familiar with this syntax
in makefiles, and its meaning is difficult to look up.

The special .IGNORE target causes make(1) to ignore all all non-zero
exit from any of the commands in that target.  In this case, it's very
unlikely that 'make preview' would ever need to run more than this one
command.

1: https://github.com/apple/swift-docc/issues/387

Co-authored-by: Kyle Murray <kyle_murray@apple.com>