Include images as prerequisites.
An alternate approach would be to include everything in the TSPL.docc
bundle:

    .../TSPL.doccarchive: Sources/TSPL/TSPL.docc/*
    .../TSPL.doccarchive: Sources/TSPL/TSPL.docc/*/*

The approach that this commit takes is more specific than the
alternative, and it encodes the conventions TSPL has around files and
folders.  The specific globs tell a contributor exactly what kind of
file should be in each place, and avoids rebuilding if the modification
date of a .DS_Store file or EmptyFile.swift is changed.  However,
we may find that this specificity limits our flexibility too much.

Incorporates feedback from Franklin Schrans <fschrans@apple.com>.