Rewrite for clarity.
When written as "avoid conflicting" or "avoid conflicts", the reader can
misread "conflict" as as verb, and then have to backtrack to re-read it
as a noun.

Applying an operator is a way to access or assign a value, so it doesn't
need to be separate.  This phrasing avoids the awkward non-parallelism
in the "X and Y, and Z" construct.