Take a different approach at the rewrite.
With the previous approach, I was trying to avoid putting prose
punctuation text to code voice punctuation, but that isn't actually as
much risk of confusion as I expected.  The "for example" about functions
and variables is a detail that was needed in the old form, to explain
when you can use the literal syntax, but isn't really needed now.

Incorporates feedback from @krilnon.