For brevity, drop suggestions for naming branches.
Because branch names don't become part of the Git history, the branch
names don't really matter.  If the branch is named 'potato' but has
descriptive commit messages and a pull request that gives reviewers
context for the changes, that's not ideal but doesn't really cause that
much of a problem.