Input em-dashes with space on both sides.
This matches the rest of the em-dashes in the book.  In most cases,
because an em-dash is a clause boundary, it ends up at the end of the
line with semantic line breaks.  The previous RST-based publication
pipeline included a step to close up em-dashes, so these few exceptions
didn't cause issues.

DocC doesn't close up em-dashes, and the current instruction from
editorial is to set em-dashes open because the San Francisco font
otherwise sets things too close together.  When setting this content in
another font that doesn't have that issue, like Helvetica, we'd need to
teach DocC to close up the em-dashes throughout.