Avoid using "data type" to mean just "type".
The term "data type" is rare in the book to begin with.  I've left a
couple places where it refers to a type that stores some kind of data.
But in most places (as pointed out by Kyle during review of this PR)
we're talking about both types that do and don't have storage.