Add missing "prints" comment.
The old tested version used << for a silent expectation, so it might
have been intentional that we omitted this.  However, the commit [1]
that added this example doesn't call out any reason for omitting the
"prints" comment, so I'm guessing it was an oversight.  Either way, it's
better to be consistent -- this was the only call to print(...) that
didn't have a comment.

[1] dda456277d696cb4a258c5f386e3b39376d80bc1