Answer a question by reading the compiler's tests.
The test cases that exercise this feature include switches over a Bool
with just a true and false case.

    https://github.com/apple/swift/pull/63378

This just follows the principle from the rest of Swift that switches
must be exhaustive -- it's not a special case here -- which is probably
why the SE proposal doesn't call it out like it does for 'if'
expressions needing an 'else' clause.