Handle redirects with an AppleRef in the hash.
These appear in redirects from developer.apple.com,
where TSPL used to be hosted.  For example:

   https://developer.apple.com/library/content/documentation/Swift/Conceptual/Swift_Programming_Language/GuidedTour.html#//apple_ref/doc/uid/TP40014097-CH2-ID1

An AppleRef for a conceptual document like TSPL has three parts after
the last slash, separated by hyphens: the document identifier, the
chapter identifier, and the section identifier.  In TSPL, we assigned
section IDs so they're are unique across the entire book, which is why
this code ignores the rest of the AppleRef.

When TSPL moved from developer.apple.com to docs.swift.org, the newer
Sphinx-based publication pipeline continued to use the same section
identifiers (also referred to as "stable links") as the anchor ID for
headings, dropping the rest of AppleRef.  Because they were the same,
the mapping table in the redirect code is also the same.