Put the Swift version back in the book's title.
When building with the legacy RST-based pipeline, the version was
included as a subtitle for the ePUB version, and in the navigation
sidebar for the HTML version.  DocC doesn't support ePUB output, or this
kind of customization to the sidebar.

This approach matches how we displayed the Swift version prior to using
the RST-based pipeline, when TSPL was published on developer.apple.com
and used [Tool J] and the corresponding pipeline.  At that time, we
included the Swift version at the end of the book's title.