Clean up actor code listing.
The value returned by `endIndex` is an invalid, past-the-end index --
saving it and assuming it will become valid when adding an element isn't
really good style.  Explicitly removing the last element is clearer and
simpler anyhow.

Marking `contents` as private is a little more realistic -- that gives
the actor some internal state whose concurrency is being managed.