Integrate this branch's changes to the grammar.
The grammar used to allow `inout` on any type annotation, following the
grammar changes described in SE-0031.  However, that isn't correct --
parameter modifiers like `inout` are allowed only in a function type or
a function declaration, as part of a parameter's type.  Separating the
productions for type-annotation and parameter-type-annotation adds a
small amount of duplication to the grammar, but it removes the
overproduction.