The value is already native endian.
This shift-and-add algorithm produces a native-endian number, not a
big-endian one.

    // Given the following example:
    let abcd = fourCharacterCode(for: "ABCD")!

    // Both of the following are true:
    abcd.bigEndian == UInt32(0x41424344).bigEndian
    abcd == 0x41424344

The test in commit 0e14203b6bd1e46669b9f116ab4cc0219bfc0abd is incorrect
because compares the result of this function (native-endian) to an
integer constant that's explicitly big-endian -- but I'm running that
test on a little-endian (Apple Silicon) machine.

Co-authored-by: Jonathan Grynspan <jgrynspan@apple.com>