Lead with names that are most likely to be familiar.
I looked at discussions of closures across a number of widely used
languages, in terms of what names were used as part of that discussion.
Note that these aren't necessarily the actual name of a feature in each
language, just the terms I saw folks use when teaching this (or a
similar) concept.

* C# - lambda expressions, anonymous methods
* C++ - lambda expressions
* Java - closures, first-class functions, inline functions
* JavaScript - closures, nested functions, code blocks
* PHP - closures, anonymous functions
* Python - lambdas, nested functions
* Ruby - closures, blocks, procs, lambdas
* Visual Basic - lambda expressions, function expressions, closures