Remove and old check that's no longer useful
Before migration to DocC, the old publication pipeline ran these tests,
so this test would fail if we forgot to update the version number for a
new Swift version.  That was a useful way to help make sure this chapter
got updated.  Now that we've building TSPL with DocC, and therefore no
longer able to run the tests, there's likewise no longer any reason to
keep this test around.