Discuss if-case and for-case in Control Flow
Using a pattern with these constructs isn't specific to enums, even
though that's a place where it's nearly indispensable.  It makes more
sense to introduce the 'case' version alongside each of the plain
versions -- folks are also more likely to look in that chapter than in
the Enumerations chapter if they're trying to look this up.