Name a repeated chunk in the grammar
Several places in the grammar allow an optional minus sign followed by
an integer literal, which this simplifies.  However, there is some
argument for keeping the minus sign separate, because it isn't part of
the literal.  This is a "signed integer literal" in that it's an integer
literal that can have a minus sign, not a literal of a signed integer.